cmd/otk: rename otk externals
The `otk` externals are moved into a subdirectory of `cmd` and named by
their target (they generate things for `osbuild`).
