+++ cmd/otk/osbuild-make-fstab-stage/main.go
