+++ cmd/otk/osbuild-make-partition-mounts-devices/main.go
