+++ cmd/otk/osbuild-make-partition-stages/main.go
